Move some functions from conda/plan.py to conda/instructions.py
