+++ conda/instructions.py
@@ -5,6 +5,7 @@
 from logging import getLogger
 from os.path import isfile, join
 
+from .common.io import time_recorder
 from .core.link import UnlinkLinkTransaction
 from .core.package_cache_data import ProgressiveFetchExtract
 from .deprecations import deprecated
@@ -102,3 +103,70 @@ OP_ORDER = (
     UNLINK,
     LINK,
 )
+
+
+@time_recorder("execute_plan")
+def execute_plan(old_plan, index=None, verbose=False):  # pragma: no cover
+    plan = _update_old_plan(old_plan)
+    execute_instructions(plan, index, verbose)
+
+
+def execute_instructions(
+    plan, index=None, verbose=False, _commands=None
+):  # pragma: no cover
+    """Execute the instructions in the plan
+
+    :param plan: A list of (instruction, arg) tuples
+    :param index: The meta-data index
+    :param verbose: verbose output
+    :param _commands: (For testing only) dict mapping an instruction to executable if None
+    then the default commands will be used
+    """
+    from .base.context import context
+    from .instructions import PROGRESS_COMMANDS, commands
+    from .models.dist import Dist
+
+    if _commands is None:
+        _commands = commands
+
+    log.debug("executing plan %s", plan)
+
+    state = {"i": None, "prefix": context.root_prefix, "index": index}
+
+    for instruction, arg in plan:
+        log.debug(" %s(%r)", instruction, arg)
+
+        if state["i"] is not None and instruction in PROGRESS_COMMANDS:
+            state["i"] += 1
+            getLogger("progress.update").info((Dist(arg).dist_name, state["i"] - 1))
+        cmd = _commands[instruction]
+
+        if callable(cmd):
+            cmd(state, arg)
+
+        if (
+            state["i"] is not None
+            and instruction in PROGRESS_COMMANDS
+            and state["maxval"] == state["i"]
+        ):
+            state["i"] = None
+            getLogger("progress.stop").info(None)
+
+
+def _update_old_plan(old_plan):  # pragma: no cover
+    """
+    Update an old plan object to work with
+    `conda.instructions.execute_instructions`
+    """
+    plan = []
+    for line in old_plan:
+        if line.startswith("#"):
+            continue
+        if " " not in line:
+            from .exceptions import ArgumentError
+
+            raise ArgumentError(f"The instruction {line!r} takes at least one argument")
+
+        instruction, arg = line.split(" ", 1)
+        plan.append((instruction, arg))
+    return plan
