+++ tests/test_instructions.py
@@ -6,8 +6,7 @@ import pytest
 
 from conda import instructions
 from conda.exceptions import CondaFileIOError
-from conda.exports import execute_instructions
-from conda.instructions import commands
+from conda.instructions import commands, execute_instructions
 
 
 def test_expected_operation_order():
